# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Janeiro Tour API - Premium Brazil Travel Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /destinations:
    get:
      operationId: listDestinations
      tags: [destinations]
      summary: List all destinations
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of destinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Destination"
    post:
      operationId: createDestination
      tags: [destinations]
      summary: Create a destination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
  /destinations/{id}:
    get:
      operationId: getDestination
      tags: [destinations]
      summary: Get a destination by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Destination detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
    patch:
      operationId: updateDestination
      tags: [destinations]
      summary: Update a destination
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
    delete:
      operationId: deleteDestination
      tags: [destinations]
      summary: Delete a destination
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Blog
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    patch:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # FAQs
  /faqs:
    get:
      operationId: listFaqs
      tags: [faqs]
      summary: List FAQs
      parameters:
        - name: tourId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: FAQ list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faq"
    post:
      operationId: createFaq
      tags: [faqs]
      summary: Create a FAQ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
  /faqs/{id}:
    patch:
      operationId: updateFaq
      tags: [faqs]
      summary: Update a FAQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
    delete:
      operationId: deleteFaq
      tags: [faqs]
      summary: Delete a FAQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Reviews
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - name: tourId
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Review list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Site Settings
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResult"
        "401":
          description: Invalid credentials
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMe"
        "401":
          description: Not authenticated

  # Stats endpoint for admin dashboard
...
components:
  schemas:
...
    Destination:
      type: object
      required: [id, slug, name, country, imageUrl, description, featured, sortOrder]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        nameEs:
          type: ["string", "null"]
        namePt:
          type: ["string", "null"]
        country:
          type: string
        imageUrl:
          type: string
        heroImageUrl:
          type: ["string", "null"]
        description:
          type: string
        descriptionEs:
          type: ["string", "null"]
        descriptionPt:
          type: ["string", "null"]
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
        featured:
          type: boolean
        sortOrder:
          type: integer
    DestinationInput:
      type: object
      required: [slug, name, country, imageUrl, description]
      properties:
        slug:
          type: string
        name:
          type: string
        nameEs:
          type: string
        namePt:
          type: string
        country:
          type: string
        imageUrl:
          type: string
        heroImageUrl:
          type: string
        description:
          type: string
        descriptionEs:
          type: string
        descriptionPt:
          type: string
        seoTitle:
          type: string
        seoDescription:
          type: string
        featured:
          type: boolean
        sortOrder:
          type: integer
    DestinationUpdate:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        nameEs:
          type: string
        namePt:
          type: string
        country:
          type: string
        imageUrl:
          type: string
        heroImageUrl:
          type: string
        description:
          type: string
        descriptionEs:
          type: string
        descriptionPt:
          type: string
        seoTitle:
          type: string
        seoDescription:
          type: string
        featured:
          type: boolean
        sortOrder:
          type: integer
    BlogPost:
      type: object
      required: [id, slug, title, excerpt, content, imageUrl, author, category, readTimeMinutes, featured, published, createdAt]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        titleEs:
          type: ["string", "null"]
        titlePt:
          type: ["string", "null"]
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        author:
          type: string
        category:
          type: string
        readTimeMinutes:
          type: integer
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
        featured:
          type: boolean
        published:
          type: boolean
        createdAt:
          type: string
    BlogPostInput:
      type: object
      required: [slug, title, excerpt, content, imageUrl, author, category, readTimeMinutes]
      properties:
        slug:
          type: string
        title:
          type: string
        titleEs:
          type: string
        titlePt:
          type: string
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        author:
          type: string
        category:
          type: string
        readTimeMinutes:
          type: integer
        seoTitle:
          type: string
        seoDescription:
          type: string
        featured:
          type: boolean
        published:
          type: boolean
    BlogPostUpdate:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        titleEs:
          type: string
        titlePt:
          type: string
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        author:
          type: string
        category:
          type: string
        readTimeMinutes:
          type: integer
        seoTitle:
          type: string
        seoDescription:
          type: string
        featured:
          type: boolean
        published:
          type: boolean
    Faq:
      type: object
      required: [id, question, answer, sortOrder]
      properties:
        id:
          type: integer
        tourId:
          type: ["integer", "null"]
        question:
          type: string
        questionEs:
          type: ["string", "null"]
        questionPt:
          type: ["string", "null"]
        answer:
          type: string
        answerEs:
          type: ["string", "null"]
        answerPt:
          type: ["string", "null"]
        sortOrder:
          type: integer
    FaqInput:
      type: object
      required: [question, answer]
      properties:
        tourId:
          type: integer
        question:
          type: string
        questionEs:
          type: string
        questionPt:
          type: string
        answer:
          type: string
        answerEs:
          type: string
        answerPt:
          type: string
        sortOrder:
          type: integer
    FaqUpdate:
      type: object
      properties:
        question:
          type: string
        questionEs:
          type: string
        questionPt:
          type: string
        answer:
          type: string
        answerEs:
          type: string
        answerPt:
          type: string
        sortOrder:
          type: integer
    Review:
      type: object
      required: [id, authorName, authorCountry, rating, comment, featured, createdAt]
      properties:
        id:
          type: integer
        tourId:
          type: ["integer", "null"]
        authorName:
          type: string
        authorCountry:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        rating:
          type: integer
        comment:
          type: string
        tourName:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        featured:
          type: boolean
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [authorName, authorCountry, rating, comment]
      properties:
        tourId:
          type: integer
        authorName:
          type: string
        authorCountry:
          type: string
        authorAvatarUrl:
          type: string
        rating:
          type: integer
        comment:
          type: string
        tourName:
          type: string
        source:
          type: string
        featured:
          type: boolean
    ReviewUpdate:
      type: object
      properties:
        authorName:
          type: string
        authorCountry:
          type: string
        authorAvatarUrl:
          type: string
        rating:
          type: integer
        comment:
          type: string
        tourName:
          type: string
        source:
          type: string
        featured:
          type: boolean
...
    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminAuthResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
    AdminMe:
      type: object
      required: [username, role]
      properties:
        username:
          type: string
        role:
          type: string
...