# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Janeiro Tour API - Premium Brazil Travel Platform
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  slug: string;
  title: string;
  /** @nullable */
  titleEs?: string | null;
  /** @nullable */
  titlePt?: string | null;
  excerpt: string;
  content: string;
  imageUrl: string;
  author: string;
  category: string;
  readTimeMinutes: number;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  featured: boolean;
  published: boolean;
  createdAt: string;
}