# lib/api-zod/src/generated/types/destinationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Janeiro Tour API - Premium Brazil Travel Platform
 * OpenAPI spec version: 0.1.0
 */

export interface DestinationInput {
  slug: string;
  name: string;
  nameEs?: string;
  namePt?: string;
  country: string;
  imageUrl: string;
  heroImageUrl?: string;
  description: string;
  descriptionEs?: string;
  descriptionPt?: string;
  seoTitle?: string;
  seoDescription?: string;
  featured?: boolean;
  sortOrder?: number;
}