# lib/api-zod/src/generated/types/faqInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Janeiro Tour API - Premium Brazil Travel Platform
 * OpenAPI spec version: 0.1.0
 */

export interface FaqInput {
  tourId?: number;
  question: string;
  questionEs?: string;
  questionPt?: string;
  answer: string;
  answerEs?: string;
  answerPt?: string;
  sortOrder?: number;
}