# lib/api-zod/src/generated/types/reviewUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Janeiro Tour API - Premium Brazil Travel Platform
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewUpdate {
  authorName?: string;
  authorCountry?: string;
  authorAvatarUrl?: string;
  rating?: number;
  comment?: string;
  tourName?: string;
  source?: string;
  featured?: boolean;
}